backends: remove "device" kwarg fallback
Remove the deprecated "device" kwarg fallback. We should have had a
warning for this, but it has been deprecated and undocumented for 5
years, so hopefully no one is using it.

The "adapter" kwarg is removed completely from the Android backend since
`self.__adapter` is unconditionally written over when connecting so it
had no effect anyway.