backends/winrt/client: add retry when getting GATT services
Add a retry when we fail to get GATT services in the WinRT backend. This
seems to happen often and was intermittently reproducible when running
integration tests locally. This seems to fix the issue. During testing,
we observed 1 to 3 retries before success. So a one second interval
seems reasonable and the 10 total retries should be more than enough.